Render Search component in sub pages
